ワーニングを修正　メッセージの解消　メイン関数の実行順の見直し
◆メッセージを解消
・広い範囲の型にキャストした方が動作が安定する場面でのキャストを指定
・スペルミスをチェックして修正

◆ワーニングを修正
・すべての変数をチェックし変数の初期化を追加

◆メイン関数の実行順の見直し
・ウィンドウズOS用のコマンド番号をセットするタイミングをOS制御システム生成後に変更
    - Windowsのコマンド番号をメイン関数へのエントリ後に実行していたが、ウィンドウズOS用のシステムを生成後にそのシステムに渡してセットする方法でなければ他のプラットフォームで使用しない変数を持ち続けることになるため変更した

+++ プロジェクト/Elina_Leadプロジェクト/Project/Elina_Lead/C_Post_Effect_Draw_Manager.cpp
@@ -70,7 +70,7 @@ void C_Post_Effect_Draw_Manager::M_Set_Post_Effect_Before_UI(GAME::INSTANCE::POS
 
 
 	// ポストエフェクトを描画リストに追加
-	m_this.mpr_variable.post_effect_before_ui_list.resize(set_post_effect_slot + 1);
+	m_this.mpr_variable.post_effect_before_ui_list.resize((size_t)set_post_effect_slot + 1);
 	m_this.mpr_variable.post_effect_before_ui_list[set_post_effect_slot] = in_add_post_effect;
 
 	return;
@@ -96,7 +96,7 @@ void C_Post_Effect_Draw_Manager::M_Set_Post_Effect_After_UI(GAME::INSTANCE::POST
 
 
 	// ポストエフェクトを描画リストに追加
-	m_this.mpr_variable.post_effect_after_ui_list.resize(set_post_effect_slot + 1);
+	m_this.mpr_variable.post_effect_after_ui_list.resize((size_t)set_post_effect_slot + 1);
 	m_this.mpr_variable.post_effect_after_ui_list[set_post_effect_slot] = in_add_post_effect;
 
 	return;
